backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
cryptography==43.0.0
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
google-generativeai==0.8.0
groq==0.11.0
matplotlib==3.9.0
plotly==5.24.0
kaleido==0.2.1
pandas==2.2.0
openpyxl==3.1.5
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
aiofiles==24.1.0
sse-starlette==2.1.0
